add docs + cleanup code
